requirements.txt
sphinx==8.0.2
pydata-sphinx-theme==0.15.2
python-docx==1.1.2
ruff==0.6.1
sphinx-design==0.6.1
sphinx-simplepdf==1.6.0
pydantic==2.8.2